moved new chart to the file ListNutritionGraph
